test_requirements.txt
django-discover-runner==1.0
docker==3.5.0
